add soc submit action
