+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-sparkoncosmos/src/main/kotlin/com/microsoft/azure/cosmos/spark/actions/SparkSubmitJobAction.kt
@@ -1,0 +1,7 @@
+package com.microsoft.azure.cosmos.spark.actions
+
+import com.microsoft.azure.hdinsight.spark.run.action.SeqActions
+
+class CosmosServerlessSparkSelectAndSubmitAction : SeqActions("Actions.SelectCosmosServerlessSparkType", "Actions.SubmitSparkApplicationAction")
+
+class CosmosSparkSelectAndSubmitAction : SeqActions("Actions.SelectCosmosSparkType", "Actions.SubmitSparkApplicationAction")
